size_t

+++ Source/JavaScriptCore/b3/B3LowerToAir.cpp
@@ -159,7 +159,7 @@ public:
         for (Variable* variable : m_procedure.variables()) {
             auto addResult = m_variableToTmps.add(variable, Vector<Tmp, 1>(m_procedure.resultCount(variable->type())));
             ASSERT(addResult.isNewEntry);
-            for (unsigned i = 0; i < m_procedure.resultCount(variable->type()); ++i)
+            for (size_t i = 0; i < m_procedure.resultCount(variable->type()); ++i)
                 addResult.iterator->value[i] = tmpForType(m_procedure.typeAtOffset(variable->type(), i));
         }
